move package `inmemory` as subpackgage of storage
