+++ pkg/storage/inmemory/in_memory_repository.go
@@ -1,4 +1,4 @@
-package storage_inmemory
+package inmemory
 
 import (
 	"context"
